python-language/django_rest/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
Pygments==2.19.2
sqlparse==0.5.5
